Change tmp and stop deleting files so fast
Switch to node's definition of a temporary directory so we don't have to
worry about it.

Also stop deleting images as the message gets sent to telegram as this
can cause a race condition on slow download and cause the image to never
be sent. It's not perfect but it works.